handler.py: Improved and tested node failure
Master node can correctly detect failure reported by standby node,
active standby as new active node, and pick a free node to be
assigned as a standby if available. Master node sends list of
active nodes and its own details in a file to the standby nodes,
and the same for the active nodes.